Add solves for Week 4. Character '_' removed from names of packages.
