+++ src/Week03/services/impl/AdminServiceImpl.java
@@ -1,0 +1,17 @@
+package Week03.services.impl;
+
+import Week03.entities.Admin;
+import Week03.entities.Client;
+import Week03.services.AdminService;
+
+public class AdminServiceImpl implements AdminService {
+    @Override
+    public void ban(Admin admin, Client client) {
+        admin.ban(client);
+    }
+
+    @Override
+    public void unban(Admin admin, Client client) {
+        admin.unban(client);
+    }
+}
