+++ src/Week04/entities/UserStatus.java
@@ -1,0 +1,7 @@
+package Week04.entities;
+
+public enum UserStatus {
+    ACTIVE,
+    DELETED
+}
+
