--- src/Week_03/entities/AbstractUser.java
@@ -1,47 +1,0 @@
-package Week_03.entities;
-
-public abstract class AbstractUser {
-    private String firstName;
-    private String lastName;
-    private String middleName;
-    private String phone;
-
-    AbstractUser(String firstName, String lastName, String middleName, String phone) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.middleName = middleName;
-        this.phone = phone;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getMiddleName() {
-        return middleName;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public void setMiddleName(String middleName) {
-        this.middleName = middleName;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-}
