--- src/Week_03/entities/Book.java
@@ -1,35 +1,0 @@
-package Week_03.entities;
-
-public class Book {
-    private String name;
-    private String author;
-
-    public Book(String name, String author) {
-        this.name = name;
-        this.author = author;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return "Book{" +
-                "name='" + name + '\'' +
-                ", author='" + author + '\'' +
-                '}';
-    }
-}
